devops-app/pyproject.toml
[tool.poetry]
name = "langgraph-multiagent-app"
version = "0.1.0"
description = "Self-hosted multi-agent LangGraph application"
package-mode=false
authors = ["Your Name <you@example.com>"]
readme = "README.md"

# Runtime dependencies (main group)
[tool.poetry.dependencies]
python = ">=3.12.3, <3.13"
langgraph = "^0.5.3"
langchain = "^0.3.26"
langchain-core = "^0.3.69"
langchain-mcp-adapters = "^0.1.9"
python-dotenv = "^1.1.1"
pydantic = "^2.11.7"
PyGithub = "^2.6.1"
psycopg-binary = "^3.2.9"
langchain-google-genai="^2.1.8"
fastapi = "^0.115.14"
streamlit = "^1.47.0"
langchain_nvidia_ai_endpoints = "^0.3.12"
uvicorn = {extras = ["standard"], version = "^0.35.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
psycopg = {extras = ["binary", "pool"], version = "^3.2.9"}
langgraph-checkpoint-postgres = "^2.0.23"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
